quick fix : Unnecessary unboxing

+++ src/main/java/logisticspipes/gui/orderer/GuiOrderer.java
@@ -162,7 +162,7 @@ public abstract class GuiOrderer extends LogisticsBaseGuiScreen implements IItem
         // Enchantment? Enchantment!
         Map<Integer, Integer> enchantIdLvlMap = EnchantmentHelper.getEnchantments(item.unsafeMakeNormalStack(1));
         for (Entry<Integer, Integer> e : enchantIdLvlMap.entrySet()) {
-            if (e.getKey().intValue() < Enchantment.enchantmentsList.length
+            if (e.getKey() < Enchantment.enchantmentsList.length
                     && Enchantment.enchantmentsList[e.getKey()] != null) {
                 String enchantname = Enchantment.enchantmentsList[e.getKey()].getTranslatedName(e.getValue());
                 if (enchantname != null) {
