+++ src/main/java/logisticspipes/gui/popup/RequestMonitorPopup.java
@@ -332,7 +332,7 @@ public class RequestMonitorPopup extends SubGuiScreen {
         }
         renderLinkedOrderListLines(list, left, top + 17);
         for (Float progress : list.getProgresses()) {
-            int pos = (int) (29.0F * progress.floatValue());
+            int pos = (int) (29.0F * progress);
             drawProgressPoint(left + 8, top + pos, 0xff00ff00);
         }
         GL11.glEnable(GL11.GL_TEXTURE_2D);
@@ -433,7 +433,7 @@ public class RequestMonitorPopup extends SubGuiScreen {
         }
         renderLinkedOrderListLines(list, innerLeftSide - mapX + 102, innerTopSide - mapY - 180);
         for (Float progress : list.getProgresses()) {
-            int pos = (int) (29.0F * progress.floatValue());
+            int pos = (int) (29.0F * progress);
             drawProgressPoint(innerLeftSide - mapX + 110, innerTopSide - mapY - 197 + pos, 0xff00ff00);
         }
         GL11.glEnable(GL11.GL_TEXTURE_2D);
