+++ src/main/java/logisticspipes/proxy/opencomputers/asm/BaseWrapperClass.java
@@ -94,7 +94,7 @@ public abstract class BaseWrapperClass extends AbstractValue {
             if (help.length() != 0) {
                 command.append("\n");
             }
-            int number = num.intValue();
+            int number = num;
             if (number < 10) {
                 command.append(" ");
             }
